tgbot/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
netifaces==0.10.6
psycopg2==2.9.9
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
wsproto==1.2.0
yarl==1.9.4
